requirements.txt
astroid==2.2.5
Django==2.2.4
django-admin==1.3.2
django-excel-response2==2.0.8
django-filter==2.2.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-six==1.0.4
djangorestframework==3.10.2
isort==4.3.21
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
mysqlclient==1.4.4
Pillow==6.1.0
pylint==2.3.1
PyMySQL==0.9.3
pytz==2019.2
screen==1.0.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
virtualenv==16.7.3
wrapt==1.11.2
xlwt==1.3.0
